+++ libyul/backends/evm/EthAssemblyAdapter.h
@@ -50,6 +50,7 @@ public:
 	size_t newLabelId() override;
 	size_t namedLabel(std::string const& _name, size_t _params, size_t _returns, std::optional<size_t> _sourceID) override;
 	void appendLinkerSymbol(std::string const& _linkerSymbol) override;
+	void appendDeployTimeAddress() override;
 	void appendVerbatim(bytes _data, size_t _arguments, size_t _returnVariables) override;
 	void appendJump(int _stackDiffAfter, JumpType _jumpType) override;
 	void appendJumpTo(LabelID _labelId, int _stackDiffAfter, JumpType _jumpType) override;