+++ libyul/backends/evm/NoOutputAssembly.h
@@ -59,6 +59,7 @@ public:
 	LabelID newLabelId() override;
 	LabelID namedLabel(std::string const& _name, size_t _params, size_t _returns, std::optional<size_t> _sourceID) override;
 	void appendLinkerSymbol(std::string const& _name) override;
+	void appendDeployTimeAddress() override;
 	void appendVerbatim(bytes _data, size_t _arguments, size_t _returnVariables) override;
 
 	void appendJump(int _stackDiffAfter, JumpType _jumpType) override;